Dayone patche!
Fixed a few game breaking bugs. namely one where right clicking on a stone slab with a stick would crash the game. May have forgotten that item attributes could be null.
Also added a creative tool for setting the amount of stone in a slab. I've got a ball of diiirt!