Updated to v1.3.2
Path Notes!
Added: A rubble storage system for storing and converting stone, gravel, and sand.
Added: Popup info on the Rubble Storage System to display contents at the top of the screen. Haven't added to slabs yet because.... well have you looked at that code, *shudder*
Fixed: A bug that caused clients to dc from a server when placing multi-blocks.
Fixed: A rendering issue with the tool molds not showing up right on the toolmold rack.